+++ libs/client/ui/pages/src/lib/profile/friends/index.ts
@@ -1,2 +1,3 @@
 export * from './friend-list/friend-list';
 export * from './request-list/request-list';
+export * from './add-friend-modal/add-friend-modal';
